dags/requirements.txt
pandas
mysql-connector-python==3.29.2
cassandra-driver
pyspark
apache-airflow